+++ src/ts/markdown/mediaRender.ts
@@ -1,0 +1,10 @@
+const videoRender = () => {
+    const type = [];
+};
+const audioRender = () => {
+    const type = [];
+};
+export const mediaRender = () => {
+    videoRender();
+    audioRender();
+};