+++ CHANGELOG.md
@@ -8,11 +8,15 @@
 
 ### TODO
 
-* [119](https://github.com/b3log/vditor/issues/119) file upload 
+* [121](https://github.com/b3log/vditor/issues/121) 所见即所得 `feature`
 * [120](https://github.com/b3log/vditor/issues/120) markdown-it 切换为 lute `enhancement`
+* [119](https://github.com/b3log/vditor/issues/119) file upload 
 * [117](https://github.com/b3log/vditor/issues/117) New Feature : support youtube `feature`
 * [112](https://github.com/b3log/vditor/issues/112) :xxx: 需根据对应的 emoji 渲染为 emoji，而非 :xxx: `feature`
-* WYSIWYG
+
+### v1.8.0 / 2019-09-03
+
+* [122](https://github.com/b3log/vditor/issues/122) 添加 emojiRender 和 highlightRender 接口 `feature`
 
 ### v1.7.25 / 2019-08-29